Rename subcommand "server" to "service"
