+++ cmd/commands/service/identity/loading_test.go
