+++ core/service/factory.go
@@ -23,7 +23,7 @@ import (
 
 	log "github.com/cihub/seelog"
 	"github.com/ethereum/go-ethereum/accounts/keystore"
-	identity_handler "github.com/mysterium/node/cmd/commands/server/identity"
+	identity_handler "github.com/mysterium/node/cmd/commands/service/identity"
 	"github.com/mysterium/node/communication"
 	nats_dialog "github.com/mysterium/node/communication/nats/dialog"
 	nats_discovery "github.com/mysterium/node/communication/nats/discovery"
